+++ CHANGELOG.md
@@ -10,6 +10,7 @@
 ### Features
 
 * include documents in npm package for ai agents ([#960](https://github.com/WeBankFinTech/fes-design/issues/960)) ([3cc476c](https://github.com/WeBankFinTech/fes-design/commit/3cc476c81fb652c7fc3887884d99b46a02de0b74))
+* feat(date-picker): 支持精确到小时和分钟 ([#913](https://github.com/WeBankFinTech/fes-design/issues/913))